feat: escape key cancels edit

+++ client/src/components/TodoItem/ItemEdit/ItemEdit.js
@@ -32,13 +32,14 @@ const ItemEdit = ({ todo }) => {
     todo.finishEdit()
   }
 
-  const handleEditCancel = () => {
-    if (!todo.id) {
-      todo.delete()
-      todo.todoListStore.resetItems()
-    } else {
-      todo.finishEdit()
-    }
+  const handleEditCancel = (event) => {
+    if (!event.key || event.key === 'Escape')
+      if (!todo.id) {
+        todo.delete()
+        todo.todoListStore.resetItems()
+      } else {
+        todo.finishEdit()
+      }
   }
 
   return (
@@ -49,10 +50,11 @@ const ItemEdit = ({ todo }) => {
           'h-8 flex-auto rounded-md border-2 border-primary px-2 text-sm',
           'outline-none transition-all duration-300 focus:shadow-md focus:shadow-primary/25',
         )}
-        onChange={handleTextInput}
         placeholder="I need to..."
         type="text"
         value={inputText}
+        onChange={handleTextInput}
+        onKeyUp={handleEditCancel}
       />
       <div className="flex items-center gap-2">
         <IconButton
